+++ core/database/src/main/java/com/evg/database/domain/model/TestTypeDBO.kt
@@ -16,16 +16,16 @@ open class ReadyTestTypeDBO : EmbeddedRealmObject {
     var type: String = ""
     var description: String = ""
     var score: Int = 0
-    var createdAt: Int = 0
+    var createdAt: Long = 0
 }
 
 open class LoadingTestTypeDBO : EmbeddedRealmObject {
     var type: String = ""
     var queue: Int = 0
     var progress: Int = 0
-    var createdAt: Int = 0
+    var createdAt: Long = 0
 }
 
 open class ErrorTestTypeDBO : EmbeddedRealmObject {
-    var createdAt: Int = 0
+    var createdAt: Long = 0
 }
